chore(deps): Remove `@guardian/private-infrastructure-config`
This dependency is deprecated. Rather than migrate to the replacement,
we can remove the dependency entirely!

We were using the library to build an IAM policy statement to SSM parameters within an account.
The account in question is the same one that the CFN stack is deployed into,
therefore we can use the CFN pseudo parameter `AWS::AccountId`.

Also switch to the `AWS::Region` pseudo parameter in favour of `props.env?.region` which was `undefined` in the tests.

See also:
  - https://github.com/guardian/private-infrastructure-config/pull/201
  - https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/pseudo-parameter-reference.html#available-pseudo-parameters